+++ src/libxfuse/file_extent_list.rs
@@ -72,6 +72,61 @@ impl<R: BufRead + Reader + Seek> File<R> for FileExtentList {
         }
     }
 
+    fn lseek(&mut self, _buf_reader: &mut R, offset: u64, whence: i32) -> Result<u64, i32> {
+        let sb = SUPERBLOCK.get().unwrap();
+
+        let dblock = offset >> sb.sb_blocklog;
+        match self.bmx.partition_point(|entry| entry.br_startoff <= dblock) {
+            0 => {
+                // A hole at the beginning of the file
+                if whence == libc::SEEK_HOLE {
+                    Ok(offset)
+                } else {
+                    self.bmx.first()
+                        .map(|b| b.br_startoff << sb.sb_blocklog)
+                        .ok_or(libc::ENXIO)
+                }
+            },
+            i => {
+                let cur_entry = &self.bmx[i - 1];
+                let br_end = cur_entry.br_startoff + cur_entry.br_blockcount;
+                if dblock < br_end {
+                    // In a data region
+                    if whence == libc::SEEK_HOLE {
+                        // Scan for the next hole
+                        for j in (i - 1)..self.bmx.len() - 1 {
+                            let before = &self.bmx[j];
+                            let after = &self.bmx[j + 1];
+                            let br_end = before.br_startoff + before.br_blockcount;
+                            if after.br_startoff > br_end {
+                                return Ok(br_end << sb.sb_blocklog);
+                            }
+                        }
+                        // Reached EOF without finding another hole.  Return the virtual hole at
+                        // EOF
+                        let entry = self.bmx.last().unwrap();
+                        let br_end = entry.br_startoff + entry.br_blockcount;
+                        Ok(br_end << sb.sb_blocklog)
+                    } else {
+                        Ok(offset)
+                    }
+                } else {
+                    // In a hole
+                    if whence == libc::SEEK_HOLE {
+                        Ok(offset)
+                    } else {
+                        match self.bmx.get(i) {
+                            Some(next_entry) => {
+                                Ok(next_entry.br_startoff << sb.sb_blocklog)
+                            },
+                            None => Err(libc::ENXIO)
+                        }
+                    }
+                }
+            }
+        }
+    }
+
     fn size(&self) -> XfsFsize {
         self.size
     }